Refatorando todo o código e organizando views e controllers

+++ src/main/java/br/com/model/services/AlunoService.java
@@ -21,6 +21,9 @@ public class AlunoService {
     }
 
     public void deleteById(Integer id) {
+        if (id == null) {
+            throw new IllegalArgumentException("id cannot be null");
+        }
         dao.deleteById(id);
     }
